Microsoft-Azure/Serverless-API-deployment/requirements.txt

azure-functions
azure-cosmos
requests
